// generated/prisma/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  OWNER: 'OWNER',
  CASHIER: 'CASHIER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TransactionStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const PaymentMethod = {
  CASH: 'CASH',
  QRIS: 'QRIS',
  TRANSFER: 'TRANSFER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const DiscountType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED'
} as const

export type DiscountType = (typeof DiscountType)[keyof typeof DiscountType]


export const StockAdjustmentType = {
  ADD: 'ADD',
  SUBTRACT: 'SUBTRACT'
} as const

export type StockAdjustmentType = (typeof StockAdjustmentType)[keyof typeof StockAdjustmentType]


export const TillStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED'
} as const

export type TillStatus = (typeof TillStatus)[keyof typeof TillStatus]